requirements.txt
asgiref==3.7.2
Django==5.0.2
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
phonenumbers==8.13.29
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
